Fix a few MSVC type conversion warnings

+++ src/client/TraceLog.cpp
@@ -26,7 +26,7 @@
 #include <iomanip>
 #include <iostream>
 
-long TraceLog::loc = 0;
+size_t TraceLog::loc = 0;
 
 TraceLog::TraceLog()
     : enabled(option(SRCML_COMMAND_VERBOSE)) {
